removing the file ebosComp.hh
